Add files via upload
ran on small dataset due to kernel issues on notebook (kept dying).
ran model on flattened and convoluted input Rgb data.
Fitted models on 2 and 3 hidden layers, with different map sizes. 
result - Model mse is good, but fit is not very good as models resorts to same predictions when new data comes in.